@@ -7,49 +7,48 @@
   "license": "MIT",
   "scripts": {
     "dev": "next dev -p ${PORT:=3003}",
+    "seed-local": "LAGOON_API_DB_HOST=localhost node api-db-seeding/index.js",
+    "seed": "node api-db-seeding/index.js",
     "build": "next build",
     "build-profile": "next build --profile",
-    "start": "next start -p ${PORT:=5000}",
+    "start": "next start -p ${PORT:=3000}",
     "storybook": "start-storybook -s ./public -p 6006 --ci",
     "build-storybook": "build-storybook -s ./public",
     "serve-storybook": "storybook-server -s ./public",
     "chromatic": "npx chromatic --allow-console-errors"
   },
   "dependencies": {
-    "@apollo/client": "^3.3.19",
-    "@apollo/link-error": "^2.0.0-beta.3",
+    "@apollo/client": "^3.5.10",
+    "@apollo/react-hooks": "^4.0.0",
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
     "@n8tb1t/use-scroll-position": "^2.0.3",
     "@react-keycloak/ssr": "^3.3.0",
     "@zeit/next-css": "^1.0.1",
-    "apollo-link": "^1.2.3",
-    "apollo-link-error": "^1.1.1",
-    "apollo-link-http": "^1.5.5",
-    "apollo-link-ws": "^1.0.10",
     "d3": "^5.15.0",
     "date-fns": "^2.9.0",
     "dayjs": "^1.10.7",
     "dotenv-extended": "^2.2.0",
     "express": "^4.16.4",
     "git-url-parse": "^11.1.2",
-    "graphql": "^14.5.8",
+    "graphql": "^16.3.0",
     "graphql-tag": "^2.10.0",
+    "graphql-ws": "^5.7.0",
     "isomorphic-unfetch": "^2.1.1",
     "js-cookie": "^3.0.1",
     "keycloak-js": "14.0.0",
     "moment": "^2.22.2",
     "moment-duration-format": "^2.2.2",
-    "next": "^10.2.2",
+    "next": "^12.1",
     "nookies": "^2.5.2",
     "nprogress": "^0.2.0",
     "object-hash": "^2.2.0",
     "postcss-nested": "^3.0.0",
     "ramda": "^0.25.0",
-    "react": "17.0.0",
+    "react": "^18.0.0",
     "react-beautiful-dnd": "^13.0.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "17.0.0",
+    "react-dom": "^18.0.0",
     "react-hexgrid": "^1.0.3",
     "react-highlight-words": "^0.14.0",
     "react-modal": "^3.8.1",
@@ -60,7 +59,6 @@
     "resize-observer-polyfill": "^1.5.1",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.0.3",
-    "subscriptions-transport-ws": "^0.9.19",
     "svg-inline-loader": "^0.8.0",
     "typescript": "^4.3.5",
     "webpack": "^5.0.0-rc.6"
@@ -78,12 +76,13 @@
     "@storybook/core": "^6.2.9",
     "@storybook/preset-typescript": "*",
     "@storybook/react": "^6.2.9",
-    "apollo-link-schema": "^1.2.4",
     "babel-loader": "^8.0.6",
     "babel-plugin-macros": "^2.7.1",
+    "bcryptjs": "^2.4.3",
     "chromatic": "^5.10.0-canary.0",
     "faker": "^5.5.3",
     "graphql-tools": "^4.0.6",
+    "mysql2": "^2.3.3",
     "node-sass": "^6.0.0",
     "prop-types": "^15.7.2",
     "react-is": "^16.12.0",
