@@ -8,9 +8,9 @@ const config = setConfig({
     LAGOON_UI_ICON: '%3Csvg%20id%3D%22Layer_11%22%20data-name%3D%22Layer%2011%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%2030%2030%22%3E%3Cdefs%3E%3Cstyle%3E.cls-1%7Bfill%3A%23231f20%3B%7D%3C%2Fstyle%3E%3C%2Fdefs%3E%3Ctitle%3Eamazee-io%3C%2Ftitle%3E%3Cpath%20class%3D%22cls-1%22%20d%3D%22M15%2C0A15%2C15%2C0%2C1%2C0%2C30%2C15%2C15%2C15%2C0%2C0%2C0%2C15%2C0ZM9.8%2C18.36a.79.79%2C0%2C0%2C1-.79.79H8.82A.79.79%2C0%2C0%2C1%2C8%2C18.36V11.64a.79.79%2C0%2C0%2C1%2C.79-.79H9a.79.79%2C0%2C0%2C1%2C.79.79Zm6.93-9.84L13.19%2C21.74a.51.51%2C0%2C0%2C1-.48.37h-.32a.5.5%2C0%2C0%2C1-.48-.63L15.45%2C8.26a.5.5%2C0%2C0%2C1%2C.48-.36h.32A.49.49%2C0%2C0%2C1%2C16.73%2C8.52Zm6.69%2C8.16c0%2C2.35-1.55%2C2.68-2.46%2C2.68h-.19c-.91%2C0-2.46-.33-2.46-2.68V13.32c0-2.35%2C1.55-2.68%2C2.46-2.68H21c.91%2C0%2C2.46.33%2C2.46%2C2.68Zm-1.77-3.45v3.54a.82.82%2C0%2C0%2C1-.78.87.83.83%2C0%2C0%2C1-.79-.87V13.23a.83.83%2C0%2C0%2C1%2C.79-.87A.82.82%2C0%2C0%2C1%2C21.65%2C13.23Z%22%2F%3E%3C%2Fsvg%3E',
     LAGOON_UI_DEPLOYMENTS_LIMIT: 10,
     LAGOON_UI_DEPLOYMENTS_LIMIT_MESSAGE: "Custom message for deployments limit - Contact service desk",
-    LAGOON_UI_TASKS_LIMIT: 25,
+    LAGOON_UI_TASKS_LIMIT: 0,
     LAGOON_UI_BACKUPS_LIMIT: 25,
-    LAGOON_VERSION: "v0.99",
+    LAGOON_VERSION: "v2.99-storybook",
     LAGOON_UI_TASK_BLOCKLIST: [],
     FAST_REFRESH: true
   },
