@@ -1,18 +1,30 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import getConfig from 'next/config';
-import { ApolloClient, InMemoryCache } from '@apollo/client';
-import { ApolloProvider as ApolloHooksProvider } from '@apollo/client';
-import { ApolloProvider } from '@apollo/client';
-import { ApolloLink, HttpLink } from '@apollo/client';
-
-import { WebSocketLink } from '@apollo/client/link/ws';
-import { onError } from '@apollo/link-error';
+import {
+  ApolloClient,
+  ApolloProvider,
+  InMemoryCache,
+  split,
+  HttpLink,
+  createHttpLink,
+  from
+} from '@apollo/client';
+import { GraphQLWsLink } from '@apollo/client/link/subscriptions';
+import { createClient } from 'graphql-ws';
+import { setContext } from '@apollo/client/link/context';
+import { onError } from "@apollo/client/link/error"
 import { getMainDefinition } from '@apollo/client/utilities';
+import { AuthContext } from 'lib/KeycloakProvider';
 
 
-// import { AuthContext } from 'lib/KeycloakProvider';
-import { AuthContext } from 'lib/Authenticator';
-import ErrorPage from 'pages/_error.js';
+// WebSocketLink Deprecated: https://www.apollographql.com/docs/react/api/link/apollo-link-ws/
+//import { WebSocketLink } from '@apollo/client/link/ws';
+
+// Recommmended to now use the newer graphql-ws library with the accompanying GraphQLWsLink.
+////
+// Subscription library change
+// @TODO: We need to update the api to use this library server-side also
+////
 
 const { publicRuntimeConfig } = getConfig();
 
@@ -20,68 +32,64 @@ const ApiConnection = ({ children }) => {
   return (
     <AuthContext.Consumer>
       {auth => {
-        if (!auth.authenticated) {
+        if (!auth.authenticated) return;
+
+        const getAuthorizationHeader = () => {
+          return {
+            "Access-Control-Allow-Origin": "*",
+            authorization: auth.authenticated ? `Bearer ${auth.apiToken}` : "",
+          };
+        }
+        
+        const authLink = setContext((_, { headers }) => {
+          return  {
+            ...headers,
+            ...getAuthorizationHeader(),
+          }
+        });
+
+        const httpLink = new HttpLink({
+          uri: publicRuntimeConfig.GRAPHQL_API,
+          headers: {
+            ...getAuthorizationHeader(),
+          }
+        });
+
+        const wsLink = new GraphQLWsLink(createClient({
+          url: publicRuntimeConfig.GRAPHQL_API.replace(/https/, 'wss').replace(/http/,'ws'),
+          connectionParams: {
+            authToken: auth.authenticated && auth.apiToken
+          },
+        }));
+
+        const errorLink = onError(({graphQLErrors, networkError}) => {
+          if (graphQLErrors) {
+            graphQLErrors.forEach(({message, locations, path}) => console.log(
+              `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`,
+              ),
+            )
+          }
+          if (networkError) {
+            console.log('[NetworkError]', networkError)
+          } 
+        });
+
+        const splitLink = split(({query})=>{
+          const definition =  getMainDefinition(query);
           return (
-            <ErrorPage
-              statusCode={401}
-              title="Login Required"
-              errorMessage="Please wait while we log you in..."
-            />
+            definition.kind === 'OperationDefinition' && definition.operation === 'subscription' 
           );
-        }
+        }, wsLink, httpLink);
 
-      const httpLink = new HttpLink({
-        uri: publicRuntimeConfig.GRAPHQL_API,
-        headers: {
-          authorization: `Bearer ${auth.apiToken}`
-        }
-      });
-
-        const HttpWebsocketLink = () => {
-          const wsLink = new WebSocketLink({
-            uri: publicRuntimeConfig.GRAPHQL_API.replace(/https/, 'wss').replace(
-              /http/,
-              'ws'
-            ),
-            options: {
-              reconnect: true,
-              connectionParams: {
-                authToken: auth.apiToken
-              }
-            }
-          });
-
-          return ApolloLink.split(
-            ({ query }) => {
-              const { kind, operation } = getMainDefinition(query);
-              return (
-                kind === 'OperationDefinition' && operation === 'subscription'
-              );
-            },
-            wsLink,
-            httpLink
-          );
-        };
 
         const client = new ApolloClient({
-          link: ApolloLink.from([
-            onError(({ graphQLErrors, networkError }) => {
-              if (graphQLErrors)
-                graphQLErrors.map(({ message, locations, path }) =>
-                  console.log(
-                    `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`
-                  )
-                );
-              if (networkError) console.log('[Network error]', networkError);
-            }),
-            // Disable websockets when rendering server side.
-            process.browser ? HttpWebsocketLink() : httpLink
-          ]),
+          ssrMode: typeof window === "undefined",
+          link: from([authLink, errorLink, splitLink]),
           cache: new InMemoryCache(),
           defaultOptions: {
             watchQuery: {
-              errorPolicy: "all",
-              fetchPolicy: "network-only",
+              errorPolicy: "ignore",
+              fetchPolicy: 'cache-and-network',
             },
             query: {
               errorPolicy: "all",
@@ -93,7 +101,7 @@ const ApiConnection = ({ children }) => {
           },
         });
 
-        return <ApolloProvider client={client}><ApolloHooksProvider client={client}>{children}</ApolloHooksProvider></ApolloProvider>;
+        return <ApolloProvider client={client}>{children}</ApolloProvider>;
       }}
     </AuthContext.Consumer>
   )
